agent_sdk/pyproject.toml
[project]
name = "agent_sdk"
version = "0.1.0"
description = "Infrastructure SDK for ADK-compatible agents: state, messaging, LLM/embedding, status reporting"
authors = [
    { name = "AI Agent Platform Team" }
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "psycopg2-binary",
    "redis",
    "google-adk"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
